Fix bad conflict resolution
