--- client/dashboard/components/activity-log/index.tsx
@@ -1,1 +1,0 @@
-export { ActivityActor } from './activity-actor';
