+++ client/dashboard/components/logs-activity/activity-actor.scss
