+++ client/dashboard/components/logs-activity/activity-actor.tsx
@@ -1,11 +1,112 @@
+import { JetpackLogo } from '@automattic/components/src/logos/jetpack-logo';
+import { WordPressLogo } from '@automattic/components/src/logos/wordpress-logo';
+import { __experimentalHStack as HStack, Icon } from '@wordpress/components';
 import { __ } from '@wordpress/i18n';
+import { commentAuthorAvatar, people, globe } from '@wordpress/icons';
 import type { SiteActivityLog } from '@automattic/api-core';
+import './activity-actor.scss';
+
+const ICON_SIZE = 24;
+
+function getActorPresentation( actor: ActivityActorType ) {
+	const { type, name, icon } = actor;
+	const actorName = name || __( 'Unknown' );
+
+	// Map known application/brand actors (v1 parity)
+	switch ( type ) {
+		case 'Application': {
+			if ( name === 'WordPress' ) {
+				return {
+					icon: (
+						<WordPressLogo
+							className="site-activity-logs__actor-icon-wordpress"
+							size={ ICON_SIZE }
+						/>
+					),
+					label: name,
+				};
+			}
+			if ( name === 'Jetpack' || name === 'Jetpack Boost' ) {
+				return {
+					icon: (
+						<JetpackLogo className="site-activity-logs__actor-icon-jetpack" size={ ICON_SIZE } />
+					),
+					label: name,
+				};
+			}
+			if ( name === 'Server' ) {
+				return {
+					icon: (
+						<Icon
+							className="site-activity-logs__actor-icon-server"
+							icon={ globe }
+							size={ ICON_SIZE }
+						/>
+					),
+					label: __( 'Server' ),
+				};
+			}
+			break;
+		}
+		case 'Multiple': {
+			return {
+				icon: (
+					<Icon
+						className="site-activity-logs__actor-icon-people"
+						icon={ people }
+						size={ ICON_SIZE }
+					/>
+				),
+				label: __( 'Multiple users' ),
+			};
+		}
+		case 'Happiness Engineer': {
+			return {
+				icon: <JetpackLogo className="site-activity-logs__actor-icon-jetpack" size={ ICON_SIZE } />,
+				label: __( 'Happiness Engineer' ),
+			};
+		}
+	}
+
+	// Default: avatar image if present; otherwise generic user icon
+	if ( icon?.url ) {
+		return {
+			icon: (
+				<img
+					className="site-activity-logs__actor-icon-avatar"
+					src={ icon.url }
+					alt={ actorName }
+					width={ ICON_SIZE }
+					height={ ICON_SIZE }
+				/>
+			),
+			label: actorName,
+		};
+	}
+
+	return {
+		icon: (
+			<Icon
+				className="site-activity-logs__actor-icon-default"
+				icon={ commentAuthorAvatar }
+				size={ ICON_SIZE }
+			/>
+		),
+		label: actorName,
+	};
+}
 
 type ActivityActorProps = {
 	actor?: SiteActivityLog[ 'actor' ];
 };
 
-export function ActivityActor( { actor }: ActivityActorProps ) {
-	const name = actor?.name?.trim();
-	return <span>{ name || __( 'Unknown' ) }</span>;
+export function ActivityActor( { actor }: { actor: ActivityActorProps } ) {
+	const { icon, label } = getActorPresentation( actor );
+
+	return (
+		<HStack spacing="2" alignment="left" className="site-activity-logs__actor">
+			{ icon }
+			<span>{ label }</span>
+		</HStack>
+	);
 }
